Remove obsolete test for updating widgets and consolidate recurrence widget tests in TestWidgetViewSet
- Deleted the TestUpdateWidget class as it was no longer needed.
- Merged the recurrence widget update test into TestWidgetViewSet, ensuring proper permission handling and response validation.
- Updated assertions to confirm the expected behavior when updating widget configurations.
